requirements/requirements.txt
dlib==19.22.0
numpy==1.19.5
opencv-python==4.5.3.20210927
Pillow==8.4.0
pandas==1.3.3
flask
requests
python-dotenv  # Needed for loading environment variables in config.py
mysql-connector-python
redis
certifi
Flask-SocketIO
eventlet         # <--- Add this for Flask-SocketIO background tasks